+++ zotify/__init__.py
@@ -27,8 +27,9 @@ from librespot.core import (
     TokenProvider as LibrespotTokenProvider,
 )
 from librespot.mercury import MercuryClient
-from librespot.metadata import EpisodeId, PlayableId, TrackId
+from librespot.metadata import EpisodeId, PlayableId, TrackId, AlbumId, ArtistId, ShowId
 from librespot.proto import Authentication_pb2 as Authentication
+from librespot.proto import Metadata_pb2 as Metadata
 from librespot.crypto import Packet
 from pkce import generate_code_verifier, get_code_challenge
 from requests import HTTPError, get, post
@@ -292,6 +293,20 @@ class ApiClient(LibrespotApiClient):
         except KeyError:
             return data
 
+    def build_request(
+        self,
+        method: str,
+        suffix: str,
+        headers: dict[str, Any] | None,
+        body: bytes | None,
+        url: str | None,
+    ):
+        if headers is not None:
+            headers["Accept-Languange"] = self.__session.language()
+        else:
+            headers = {"Accept-Language": self.__session.language()}
+        return super().build_request(method, suffix, headers, body, url)
+
     def __get_token(self) -> str:
         return (
             self.__session.tokens()
@@ -303,6 +318,73 @@ class ApiClient(LibrespotApiClient):
             )
             .access_token
         )
+
+    # TODO: Remove when fix is applied to librespot
+    def get_metadata_4_album(self, album: AlbumId) -> Metadata.Album:
+        """
+
+        :param album: AlbumId:
+
+        """
+        response = self.sendToUrl(
+            "GET",
+            "https://spclient.wg.spotify.com",
+            "/metadata/4/album/{}".format(album.hex_id()),
+            None,
+            None,
+        )
+        ApiClient.StatusCodeException.check_status(response)
+
+        body = response.content
+        if body is None:
+            raise IOError()
+        proto = Metadata.Album()
+        proto.ParseFromString(body)
+        return proto
+
+    # TODO: Remove when fix is applied to librespot
+    def get_metadata_4_artist(self, artist: ArtistId) -> Metadata.Artist:
+        """
+
+        :param artist: ArtistId:
+
+        """
+        response = self.sendToUrl(
+            "GET",
+            "https://spclient.wg.spotify.com",
+            "/metadata/4/artist/{}".format(artist.hex_id()),
+            None,
+            None,
+        )
+        ApiClient.StatusCodeException.check_status(response)
+        body = response.content
+        if body is None:
+            raise IOError()
+        proto = Metadata.Artist()
+        proto.ParseFromString(body)
+        return proto
+
+    # TODO: Remove when fix is applied to librespot
+    def get_metadata_4_show(self, show: ShowId) -> Metadata.Show:
+        """
+
+        :param show: ShowId:
+
+        """
+        response = self.sendToUrl(
+            "GET",
+            "https://spclient.wg.spotify.com",
+            "/metadata/4/show/{}".format(show.hex_id()),
+            None,
+            None,
+        )
+        ApiClient.StatusCodeException.check_status(response)
+        body = response.content
+        if body is None:
+            raise IOError()
+        proto = Metadata.Show()
+        proto.ParseFromString(body)
+        return proto
 
 
 class TokenProvider(LibrespotTokenProvider):